Using WK Item Info Injector
WK updated parts of the lesson page to use React, which breaks this script. This update lets WK Item Info Injector handle the injection.
This version is also more secured against XSS attacks, but this change was probably unnecessary because Kanji Damage is likely trustworthy, and the pages are stored on https://grenzionky.github.io
The currently allowed tags are only p, div, a, span, img, and br, which is probably not sufficient.